+++ examples/qt_test.py
@@ -6,8 +6,7 @@ This example demonstrates how to create a pywebview windows using QT (normally G
 """
 
 if __name__ == '__main__':
-    webview.config["USE_QT"] = True
+    webview.config.use_qt = True
 
     # Create a non-resizable webview window with 800x600 dimensions
     webview.create_window("Simple browser", "http://flowrl.com")
-