+++ webview/__init__.py
@@ -28,10 +28,18 @@ FOLDER_DIALOG = 20
 SAVE_DIALOG = 30
 
 
-config = {
-    "USE_QT": False,
-    "USE_WIN32": False
-}
+class Config (dict):
+    use_qt = False
+    use_win32 = False
+
+    def __getitem__(self, key):
+        return getattr(Config, key.lower())
+
+    def __setitem__(self, key, value):
+        setattr(Config, key.lower(), value)
+
+
+config = Config()
 
 _initialized = False
 _webview_ready = Event()
@@ -39,35 +47,34 @@ _webview_ready = Event()
 
 def _initialize_imports():
     global _initialized, gui
+    import_error = False
 
     if not _initialized:
         if platform.system() == "Darwin":
-            try:
-                import webview.cocoa as gui
-            except ImportError:
-                import_error = True
-            else:
-                import_error = False
+            if not config.use_qt:
+                try:
+                    import webview.cocoa as gui
+                except ImportError:
+                    logger.exception("PyObjC cannot be loaded")
+                    import_error = True
 
-            if import_error or config["USE_QT"]:
+            if import_error or config.use_qt:
                 try:
                     import webview.qt as gui
                     logger.info("Using QT")
                 except ImportError as e:
                     # Panic
-                    logger.exception("QT not found")
+                    logger.exception("QT cannot be loaded")
                     raise Exception("You must have either PyObjC (for Cocoa support) or Qt with Python bindings installed in order to use this library.")
+
         elif platform.system() == "Linux":
-            try:
-                #Try GTK first unless USE_QT flag is set
-                if not config["USE_QT"]:
+            if not config.use_qt:
+                try:
                     import webview.gtk as gui
                     logger.info("Using GTK")
-            except ImportError as e:
-                logger.exception("GTK not found")
-                import_error = True
-            else:
-                import_error = False
+                except ImportError as e:
+                    logger.exception("GTK cannot be loaded")
+                    import_error = True
 
             if import_error or config["USE_QT"]:
                 try:
@@ -76,30 +83,28 @@ def _initialize_imports():
                     logger.info("Using QT")
                 except ImportError as e:
                     # Panic
-                    logger.exception("QT not found")
+                    logger.exception("QT cannot be loaded")
                     raise Exception("You must have either QT or GTK with Python extensions installed in order to use this library.")
 
         elif platform.system() == "Windows":
-
-            try:
-                #Try .NET first unless USE_WIN32 variable is defined
-                if not config["USE_WIN32"]:
+            #Try .NET first unless use_win32 flag is set
+            if not config.use_win32:
+                try:
                     import webview.winforms as gui
                     logger.info("Using .NET")
-            except ImportError as e:
-                logger.exception("pythonnet not found")
-                import_error = True
-            else:
-                import_error = False
+                except ImportError as e:
+                    logger.exception("pythonnet cannot be loaded")
+                    import_error = True
 
-            if import_error or config["USE_WIN32"]:
+
+            if import_error or config.use_win32:
                 try:
                     # If .NET is not found, then try Win32
                     import webview.win32 as gui
                     logger.info("Using Win32")
                 except ImportError as e:
                     # Panic
-                    logger.exception(" not found")
+                    logger.exception("PyWin32 cannot be loaded")
                     raise Exception("You must have either pythonnet or pywin32 installed in order to use this library.")
         else:
             raise Exception("Unsupported platform. Only Windows, Linux and OS X are supported.")
@@ -149,14 +154,14 @@ def load_html(content, base_uri=""):
     :param base_uri: Base URI for resolving links. Default is "".
     """
     try:
-        _webview_ready.wait(50)
+        _webview_ready.wait(5)
         gui.load_html(_make_unicode(content), base_uri)
     except NameError as e:
         raise Exception("Create a web view window first, before invoking this function")
 
 
 def create_window(title, url=None, width=800, height=600,
-                  resizable=True, fullscreen=False, min_size=(200, 100), strings={}):
+                  resizable=True, fullscreen=False, min_size=(200, 100), strings={}, confirm_quit=False):
     """
     Create a web view window using a native GUI. The execution blocks after this function is invoked, so other
     program logic must be executed in a separate thread.
@@ -172,7 +177,7 @@ def create_window(title, url=None, width=800, height=600,
     """
     _initialize_imports()
     localization.update(strings)
-    gui.create_window(_make_unicode(title), _transform_url(url), width, height, resizable, fullscreen, min_size, _webview_ready)
+    gui.create_window(_make_unicode(title), _transform_url(url), width, height, resizable, fullscreen, min_size, confirm_quit, _webview_ready)
 
 
 def get_current_url():