+++ webview/cocoa.py
@@ -26,6 +26,30 @@ class BrowserView:
     app = AppKit.NSApplication.sharedApplication()
 
     class AppDelegate(AppKit.NSObject):
+        def applicationDidFinishLaunching_(self, notification):
+            BrowserView.instance.webview_ready.set()
+
+    class WindowDelegate(AppKit.NSObject):
+        def display_confirmation_dialog(self):
+            AppKit.NSApplication.sharedApplication()
+            AppKit.NSRunningApplication.currentApplication().activateWithOptions_(AppKit.NSApplicationActivateIgnoringOtherApps)
+            alert = AppKit.NSAlert.alloc().init()
+            alert.addButtonWithTitle_(localization["global.quit"])
+            alert.addButtonWithTitle_(localization["global.cancel"])
+            alert.setMessageText_(localization["global.quitConfirmation"])
+            alert.setAlertStyle_(AppKit.NSWarningAlertStyle)
+
+            if alert.runModal() == AppKit.NSAlertFirstButtonReturn:
+                return True
+            else:
+                return False
+
+        def windowShouldClose_(self, notification):
+            if not _confirm_quit or self.display_confirmation_dialog():
+                return Foundation.YES
+            else:
+                return Foundation.NO
+
         def windowWillClose_(self, notification):
             BrowserView.app.stop_(self)
 
@@ -33,8 +57,73 @@ class BrowserView:
         def webView_contextMenuItemsForElement_defaultMenuItems_(self, webview, element, defaultMenuItems):
             return nil
 
-        def webView_didFinishLoadForFrame_(self, webview, frame):
-            BrowserView.instance.webview_ready.set()
+        # Display a JavaScript alert panel containing the specified message
+        def webView_runJavaScriptAlertPanelWithMessage_initiatedByFrame_(self, webview, message, frame):
+            AppKit.NSRunningApplication.currentApplication().activateWithOptions_(AppKit.NSApplicationActivateIgnoringOtherApps)
+            alert = AppKit.NSAlert.alloc().init()
+            alert.setInformativeText_(message)
+            alert.runModal()
+
+        def webView_printFrameView_(self, webview, frameview):
+            """
+            This delegate method is invoked when a script or a user wants to print a webpage (e.g. using the Javascript window.print() method)
+            :param webview: the webview that sent the message
+            :param frameview: the web frame view whose contents to print
+            """
+            def printView(frameview):
+                # check if the view can handle the content without intervention by the delegate
+                can_print = frameview.documentViewShouldHandlePrint()
+
+                if can_print:
+                    # tell the view to print the content
+                    frameview.printDocumentView()
+                else:
+                    # get an NSPrintOperaion object to print the view
+                    info = AppKit.NSPrintInfo.sharedPrintInfo().copy()
+
+                    # default print settings used by Safari
+                    info.setHorizontalPagination_(AppKit.NSFitPagination)
+                    info.setHorizontallyCentered_(Foundation.NO)
+                    info.setVerticallyCentered_(Foundation.NO)
+
+                    imageableBounds = info.imageablePageBounds()
+                    paperSize = info.paperSize()
+                    if (Foundation.NSWidth(imageableBounds) > paperSize.width):
+                        imageableBounds.origin.x = 0
+                        imageableBounds.size.width = paperSize.width
+                    if (Foundation.NSHeight(imageableBounds) > paperSize.height):
+                        imageableBounds.origin.y = 0
+                        imageableBounds.size.height = paperSize.height
+
+                    info.setBottomMargin_(Foundation.NSMinY(imageableBounds))
+                    info.setTopMargin_(paperSize.height - Foundation.NSMinY(imageableBounds) - Foundation.NSHeight(imageableBounds))
+                    info.setLeftMargin_(Foundation.NSMinX(imageableBounds))
+                    info.setRightMargin_(paperSize.width - Foundation.NSMinX(imageableBounds) - Foundation.NSWidth(imageableBounds))
+
+                    # show the print panel
+                    print_op = frameview.printOperationWithPrintInfo_(info)
+                    print_op.runOperation()
+
+            PyObjCTools.AppHelper.callAfter(printView, frameview)
+
+        # WebPolicyDelegate method, invoked when a navigation decision needs to be made
+        def webView_decidePolicyForNavigationAction_request_frame_decisionListener_(self, webview, action, request, frame, listener):
+            # The event that might have triggered the navigation
+            event = AppKit.NSApp.currentEvent()
+            action_type = action['WebActionNavigationTypeKey'] 
+
+            """ Disable back navigation on pressing the Delete key: """
+            # Check if the requested navigation action is Back/Forward
+            if action_type == WebKit.WebNavigationTypeBackForward:
+                # Check if the event is a Delete key press (keyCode = 51)
+                if event and event.type() == AppKit.NSKeyDown and event.keyCode() == 51:
+                    # If so, ignore the request and return
+                    listener.ignore()
+                    return
+
+            # Normal navigation, allow
+            listener.use()
+
 
     class WebKitHost(WebKit.WebView):
         def performKeyEquivalent_(self, theEvent):
@@ -98,10 +187,13 @@ class BrowserView:
         self.window.setContentView_(self.webkit)
 
         self._browserDelegate = BrowserView.BrowserDelegate.alloc().init()
+        self._windowDelegate = BrowserView.WindowDelegate.alloc().init()
         self._appDelegate = BrowserView.AppDelegate.alloc().init()
         self.webkit.setUIDelegate_(self._browserDelegate)
         self.webkit.setFrameLoadDelegate_(self._browserDelegate)
-        self.window.setDelegate_(self._appDelegate)
+        self.webkit.setPolicyDelegate_(self._browserDelegate)
+        self.window.setDelegate_(self._windowDelegate)
+        BrowserView.app.setDelegate_(self._appDelegate)
 
         self.load_url(url)
 
@@ -282,7 +374,7 @@ class BrowserView:
         return val
 
 
-def create_window(title, url, width, height, resizable, fullscreen, min_size, ready_event):
+def create_window(title, url, width, height, resizable, fullscreen, min_size, confirm_quit, ready_event):
     """
     Create a WebView window using Cocoa on Mac.
     :param title: Window title
@@ -292,6 +384,9 @@ def create_window(title, url, width, height, resizable, fullscreen, min_size, re
     :param resizable True if window can be resized, False otherwise
     :return:
     """
+    global _confirm_quit
+    _confirm_quit = confirm_quit
+
     browser = BrowserView(title, url, width, height, resizable, fullscreen, min_size, ready_event)
     browser.show()