+++ webview/win32.py
@@ -287,7 +287,10 @@ class BrowserView(object):
             custom_doc.SetUIHandler(self.handler)
 
 
-def create_window(title, url, width, height, resizable, fullscreen, min_size, webview_ready):
+def create_window(title, url, width, height, resizable, fullscreen, min_size, confirm_quit, webview_ready):
+    global _confirm_quit
+    _confirm_quit = confirm_quit  # not implemented
+
     set_ie_mode()
     browser_view = BrowserView(title, url, width, height, resizable, fullscreen, min_size, webview_ready)
     browser_view.show()